+++ src/arcid/components/FlightProfile/PointProfile.jsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom';
 
 import _ from 'lodash';
 import moment from 'moment';
+import R from 'ramda';
 
 import {
   Table,
@@ -16,7 +17,6 @@ import theme from '../../../theme';
 import PointProfileLegend from './PointProfileLegend';
 import ColoredFlightLevel from './ColoredFlightLevel';
 import ColorizedContent from '../../../core/components/ColorizedContent';
-import Trend from './Trend';
 
 const defaultStyles = {
   table: {
@@ -124,8 +124,6 @@ class PointProfile extends Component {
 
     const styles = Object.assign({}, defaultStyles, style);
 
-    let previousTrend;
-
     return (
       <div
         style={{
@@ -145,8 +143,8 @@ class PointProfile extends Component {
             style={styles.tableBody}
           >
             {_.map(pointProfile, (point, index) => {
-              const showTrend = previousTrend !== point.trend || point.trend !== 'CRUISE';
-              previousTrend = point.trend;
+              const center = R.pathOr('XXXX', ['airspace', 'center'], point);
+              const sector = R.pathOr('XXXX', ['airspace', 'name'], point);
 
               return (
                 <TableRow
@@ -174,19 +172,21 @@ class PointProfile extends Component {
                   <TableRowColumn
                     style={styles.tableBodyColumn}
                   >
-                    <ColorizedContent hash={_.get(point, 'airspace.center', 'XXXX')}>
-                      {_.get(point, 'airspace.center', 'XXXX')}
+                    <ColorizedContent hash={center}>
+                      {center}
+                    </ColorizedContent>
+                  </TableRowColumn>
+                  <TableRowColumn
+                    style={styles.tableBodyColumn}
+                  >
+                    <ColorizedContent hash={sector}>
+                      {sector}
                     </ColorizedContent>
                   </TableRowColumn>
                   <TableRowColumn
                     style={styles.tableBodyColumn}
                   >
                     <ColoredFlightLevel flightLevel={point.flightLevel} />
-                  </TableRowColumn>
-                  <TableRowColumn
-                    style={styles.tableBodyColumn}
-                  >
-                    {showTrend && <Trend trend={point.trend} />}
                   </TableRowColumn>
                 </TableRow>
               );